Previous fix for (Fix "hang" or "hangup" when 0 length .wav files end up in voicemail) broke playing http fetched prompts
Instead of calling DataSource::getLength(), use DataSource::peek() instead
to "pre-read" 4 bytes, to check if this is a 0 length file and should
be skipped.  Alas, apparently one cannot just call getLength() 
as an http fetch might not have returned any info yet, and so it may (or may
not) return 0 length depending on how quickly the http server responded.

Fixes XPB-1004 (Auto Attendant is not working fine - "intermittent")

--Woof!


git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@9218 a612230a-c5fa-0310-af8b-88eea846685b