Changes to OsDir getFileInfo for Windows CE 6.0 (and prior)
Prior to this, there was a Posix implementation in OsDirBase, 
After this, Posix implementation moved to OsDirLinux, and new implementation is created in OsDirWnt.

getFileInfo now is labeled as a const method.

NOTE: This breaks VXWorks.  Correct method of implementation is probably to create OsDirPosix, and have any platforms that can use posix code use this.

git-svn-id: http://scm.sipfoundry.org/rep/sipX/branches/sipXtapi@9241 a612230a-c5fa-0310-af8b-88eea846685b