Implemented XPL-204 - static initialization solution.
Read the *Init.h files for detailed description on how to fix static initialization problem if discovered and how to avoid it. From now on, it is safe to use static objects in static constructor/destructor if the instructions are followed.

Made TapiMgr a static singleton without using new, as it's not needed.

Made spSipMessageFieldProps in SipMessage a static singleton without using new. Constructor/destructor ordering problem is avoided in the way described in UtlInit.h.


git-svn-id: http://scm.sipfoundry.org/rep/sipX/branches/sipXtapi@9571 a612230a-c5fa-0310-af8b-88eea846685b