Add runtime settable codec path support to CpMediaInterfaceFactory.
To use, supply codec paths prior to instantiating a CpMediaInterfaceFactory,
as codec paths are supplied to media subsystem upon media factory construction.
An example can be seen in CpPhoneMediaInterfaceTest.

To make purify and variants happy, be sure to clear the codec paths with clearCodecPaths after you are done using any factories.

git-svn-id: http://scm.sipfoundry.org/rep/sipX/branches/sipXtapi@10153 a612230a-c5fa-0310-af8b-88eea846685b