Add new wideband input/output test.
This tests a variety of tones at a variety of sample rates, to test wideband support in flowgraphs.

Note: this test requires a short-circuit cable - male minijack-to-male minijack, so it can record some generated output, and play it back.  It also requires user intervention to properly run and interpret the test.

git-svn-id: http://scm.sipfoundry.org/rep/sipX/branches/sipXtapi@10268 a612230a-c5fa-0310-af8b-88eea846685b