Rewrite speex detection and usage code. Now add separate enable/disable for speex usage in main sipXmediaLib code (dsp routines like resampling, VAD, etc..), and in building speex codec plugin.
Usage of contrib libspeex code is also made clearer -- if you want to make
sure that contrib code is used, specify --enable-contrib-speex.  Otherwise, if
the proper version of speex is detected installed on your system, it will use
that.  If no installed speex of proper version or above is detected, it will
always configure to build and use contrib code.

Here is a summary of the related configure options to this patch:
--enable-contrib-speex -- force usage of speex contained in contrib dir over all else. (NEW)
--enable-codec-speex -- Build and install the speex codec plugin.
--enable-speex-dsp -- enable use of speex in main sipXmediaLib code (just the speex dsp routines like resampling, etc) (NEW)

CAUTION:
This does have a slight drawback in the way it is engineered -- if contrib
libspeex ends up being selected (by forcing it, or just by falling back to it), it will currently be installed if you 'make install' sipXmediaLib.
Beware of this, as, if you have an installed speex, it will overwrite your
installed speex libs and includes.

A patch forthcoming will fix this.



git-svn-id: http://scm.sipfoundry.org/rep/sipX/branches/sipXtapi@10424 a612230a-c5fa-0310-af8b-88eea846685b