API change to sipXmediaAdapterLib for specifying media interface samples per frame and sample rate.
createMediaInterface now takes both of these parameters, and it is then passed through such that the associated flowgraph is created with the specified sample rate, samples per frame.

If no sample rate is specified, the usual default of 8000Hz, and 80 samples per second is used.  Those defaults are now determined in MpFlowGraphBase.cpp

git-svn-id: http://scm.sipfoundry.org/rep/sipX/branches/sipXtapi@10492 a612230a-c5fa-0310-af8b-88eea846685b