Work around brokenness in VC6.
In MSVC6, MprRtpDispatcherActiveSsrcs::lookupRtpStream fails to compile
properly with MpRtpStream defined in MprRtpDispatcher as protected. 
VC6 erroneously specifies that it 'cannot access protected struct 
declared in class 'MprRtpDispatcher' despite the fact that 
MprRtpDispatcherActiveSrcs derives from MprRtpDispatcher, and defines 
lookupRtpStream as protected.

So, to work around this, if we're compiling with VC6, then make the
struct public (which exposes it more than we want)...

This only affects VC6 builds.



git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@11027 a612230a-c5fa-0310-af8b-88eea846685b